Implement Order Creation and Basket Clearing Logic.
Create an Order from the Basket:

Fetch the user's current basket.

Convert basket items into an order.

Save the order in the database.

Clear the Basket:

Mark the current basket as deleted.

Remove all items from the basket.

Integration:

Ensure OrderController and OrderService work with BasketService.

Add logging and error handling.

Testing:

Verify the order is created correctly.

Ensure the basket is cleared and ready for new items.

1. OrderController
Endpoint: POST /api/Order/CreateOrder

Input: OrderCreateDto (contains deliveryTime and address).

Steps:

Validate Input: Check if deliveryTime and address are valid.

Extract User ID: Get the authenticated user's ID from the claims.

Create Order: Call OrderService.CreateOrderFromBasketAsync to create an order from the user's basket.

Clear Basket: Call BasketService.ClearBasketAsync to clear the user's basket after order creation.

Return Response: Return success or error messages based on the operation.

2. OrderService
Method: CreateOrderFromBasketAsync

Steps:

Fetch Basket: Retrieve the user's active basket (where DeleteDateTime is null).

Check Basket: If the basket is empty, return null.

Create Order:

Map basket items to order items.

Set order details (e.g., deliveryTime, address, status, price).

Save Order: Add the order to the database and save changes.

Return Order DTO: Map the order to OrderDto and return it.

3. BasketService
Method: ClearBasketAsync

Steps:

Fetch Basket: Retrieve the user's active basket.

Mark as Deleted: Set DeleteDateTime to the current time.

Remove Items: Delete all items in the basket.

Create New Basket: Create a new, empty basket for the user.

Save Changes: Persist changes to the database.

4. Integration
OrderController:

Calls OrderService to create an order.

Calls BasketService to clear the basket.

OrderService:

Works with BasketService to fetch the basket and convert it into an order.

BasketService:

Clears the basket and creates a new one for the user.

5. Logging and Error Handling
Logging:

Added logging in OrderController, OrderService, and BasketService to track operations (e.g., order creation, basket clearing).

Error Handling:

Used try-catch blocks to handle exceptions.

Return appropriate HTTP status codes (e.g., 400 Bad Request, 500 Internal Server Error).

+++ DTO/DishBasketDto.cs
@@ -10,8 +10,8 @@ namespace DTO
     {
         public Guid Id { get; set; }
         public string Name { get; set; }
-        public double Price { get; set; }
-        public double TotalPrice { get; set; }
+        public decimal Price { get; set; }
+        public decimal TotalPrice { get; set; }
         public int Amount { get; set; }
         public string? Image { get; set; } 
     }
